dutyCycleEncoder, falcons turn and drive, discretize
DutyCycleEncoder was rewritten in 2025 WPILib. Updated that

Made both the drive and turn motors of each module into TalonFX since we will be using Krakens for drive and Falcons for steering.

Changed to Canandgyro

Discretized ChassisSpeeds to hopefully correct for translational skew (I'm praying to Jesus H. Christ that this works) Will test tomorrow.

Forced to utilize the WPILib units library more in 2025. Probably not a bad thing for avoiding unit conversion errors.

Changed Joystick axes in SwerveJoystick to right numbers

Did some robot.java and robotContainer.java stuff. Should be able to auto and tele.