Odometry Thread
Used the wrong method last time to sleep the thread, causing the robot to tweak out. Changed it to the right method Thread.sleep(10), robot behaved as expected, just with a higher odometry update rate. Print statements showed faster odometry updates. Tested it in the field with the 2024 robot. DID NOT IMPROVE ODOMETRY 😭