Autos Working
Things I had to fix to get auto to work:

Downloaded Pathplanner from the github instead of the microsoft store version.

Doing this got the .path files to appear in the deploy path folder. Before that, the path folder was totally devoid of .path files.

Once I got the path files to appear, the sendable chooser for the AutoBuilder still didn't show any autos. But this was a fake. The drop-down box look likes it only has one option, but if you place your cursor carefully on the dropdown and scroll, you can see that all of the autos are there.

I selected an auto that had newly appeared now that the path files were in the folder. I ran it, nothing happened. Ran it again, heard a pretty loud click in the gearbox, so it definitely wanted to move.

Realized the pathplanner constants were placeholder and I got them to the right values, driving normally now.

Figured out how to register named commands, and got that to work in auto.

Climb stuff still a big goofy